Add reCAPTCHA to the contact form
A script was added to the contact.html file to utilize Google's reCAPTCHA service to protect against spam and other types of automated abuse.
Instead of the previous submit button, now the reCAPTCHA widget serves as the submission button, as users must interact with it before submitting the form, validating their non-robot identity.
A minor change is also seen in the .iml file where an "api" library is added.