Update contact form and background color
In this commit, the contact form on the main page has been given its own unique style by adding an extra CSS class 'contactFormStyle'. This was needed to add specific padding to the top and bottom of the form for better presentation. Also, the lightBg color variable used in CSS for background color has been darkened to a new shade for improved visibility and accessibility.