Add image cropping to skills card
This commit introduces an image cropping feature to the 'keySkillCard' in index.html by wrapping the existing 'img' tag inside a new div with a class of 'image-cropper'.

Also, the styles associated with the 'image-cropper' are added to the style.css file. The styles make any image appearing within the 'image-cropper' to be rendered in a circular shape with a radius of 50%.

The update is aimed at enhancing the visual appeal of the skills section on the site.