Add GitHub icon and adjust blog images styling
A GitHub icon is added to the site's footer for improved user accessibility to the GitHub page. The error message related to a blog without an associated image file has been removed and the image styling for blog posts updated. This resolves previous ValueError occurrence in the blog template, and additionally makes the visual layout for blog posts more appealing by adjusting image cropping.