Fix progress bar width in the main index
The previous code was setting the progress bar width directly with a variable. This meant that as the variable value changed, the size of the bar reduced considering the variable as a percentage. This was leading to misrepresentation of the progress.

This commit makes the change so that the progress bar width is calculated as a difference to 100%. This ensures that progress bar represents the opposite of the variable value.