Reduce avatar size and adjust keySkillCard padding
Update the style.css to reduce the size of the avatar from 70px to 50px to make it more unobtrusive. Also made changes in .keySkillCard by defining flex properties to align items and added spacing between them. Adjusted the padding-top from 25px to 10px for an improved layout. Changes made to optimize the UI aesthetics.