Change CV field from CloudinaryField to FileField
The field for storing user CVs in the main models is updated. Previously, it was a CloudinaryField which was used for cloud storage. Due to recent changes in our app architecture and to ease local development and testing, we've switched this field to a Django FileField. The uploaded files will now be stored in a "cv" directory.